move extensions under the related subpackages
